requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
langgraph
pgvector
psycopg2
requests